"9:30am. Time sure passes fast. The tablet should arrive in the next 1-3 days. Let me chill a bit and then I will start leg sculpting.
9:50am. Let me sculpt. I want to play around at making things smooth. I really should restart the sculpting and do it like Flycat - with the minimum of effort, but I want to internalize the feel. I do not want to put in a bunch of object and adjust it by tiny amount. I should be able to sculpt things in place for any kind of position.

10:05am. I am getting it. A combination of inflate + smooth tool works very well. The upper part needs to be more taut, and the lower legs are too short, but I've managed to get over the noodle problems.

Right now it no longer looks like a 50 yo housewife, but maybe a 30 yo who hit the wall.

https://www.youtube.com/results?search_query=blender+measure

Focus me, stop listening the Ruina ost and watch this thing.

10:15am. Let me get rid of one of the legs. It is just getting too much in the way.

10:45am. I keep going back and forth on the leg, trying to get the stars to align. Forget the upper part, let me give some definition to the lower part. But even so the upper part still have that woman-who-hit-the-wall feel.

11:15am. Hmmm, I didn't do the toes, but this is good enough to pause and review.

The look still has that woman-past-her-prime feel, and to be honest, I am not sure why. If I had such perfect self awareness that I could immediatelly understand how to fix something based on knowing the problem, I would have already made good RL agents.

I keep trying to capture Flycat's feeling and am failing. Let me stop here with my own effort for now. I think once I get the tablet and start drawing, my sculpting and my drawing efforts will synergize. Let me not forget that my grasp of the human body is still quite incomplete at this stage.

11:20am. Let me watch how Devilpeace does it.

https://www.youtube.com/watch?v=axOS7n9e6to
Sculpting Leg ( Female ) Blender sculpting tutorial 2.9+ Slow and detailed

https://youtu.be/axOS7n9e6to?t=3

One thing I notice that the calves for this model are a lot bigger than mine. I know that Flycat's model is unnatural, so I've been doing the lower legs on the smaller side. But still, I do feel that the upper half is saggy even in isolation.

https://youtu.be/axOS7n9e6to?t=168

Just what am I looking at, aren't these obviously guy legs?

https://youtu.be/axOS7n9e6to?t=558

Ok, the upper ref is female.

11:35am. I skimmed the video. I don't know. None of this tells me what exactly is wrong with my own model. Let me get rid of the bottom leg.

11:40am. I just do not get it. Is the leg shape rounded in an uneven manner.

Let me watch Flycat's video.

Focus me.

https://www.youtube.com/watch?v=VzMAh66ofq0
Create a Perfect Body Base Mesh with Blender 2.92

Let me watch the first 10m of this again.

11:45am. https://youtu.be/VzMAh66ofq0?t=16

Actually right at the start, it gives me a complete reference.

Looking at this, I can immediatelly see what the problem with my own model is. If you look at the knee, it is not all the way to the right and the thigh does not just drop down. Instead the upper thigh curves outward past the crotch and then inward as it nears the knee. This gives it a feeling of tautness. Similarly, it also curves out past the waist and then inwards.

I fucked up those curves, so it is no wonder I get the feeling I am looking at saggy woman hips and legs. I went and improved the lower leg, but that did not change my perception much. The only thing I could think of is that the leg is not smooth enough, but that is not really the problem.

https://youtu.be/VzMAh66ofq0?t=29

This part where he is rotating the model is everything I need.

Sigh, I am so incompetent.

11:55am. I am taking a breather. Let me stop here for breakfast. I am going to use this as reference and see if I can improve the leg later. Tomorrow the tablet might arrive and I want to get ready.

These studies are going to become useful when it is time to draw chara portraits. Plus, it would be good to be able to showcase something."