"11am. > The parcel is expected to be delivered during the day.
Sweet, I did not expect it would be this soon. I only just now.

I'll start soon.

The order originally said between 3.11 and 5.11. I thought that might have been a lie, but it was not.

11:05am. Nevermind that for now. I am a bit excited, but let me spend my energy on sculpting instead.

11:10am. Did the initial bit. Now let me put up the reference.

11:20am. I've got the front, sides and back set up. Now I can start. Let me hash out the initial shape.

12:20pm. I am reviewing. I got the shape down, but it has a very stocky feel. Not necessarily old housewife stocky, but something like that bug judo girl from Hagure Idol stocky.

https://youtu.be/VzMAh66ofq0?t=27

When you look at Flycat's legs, they feel very lithe.

12:40pm. Let stop here for the morning.

I think the upper half of the lgs is actually passable. It is not as good at Flycat's and feels a tad bulky, but it is fine.

The feeling of stockiness only really hits me once I zoom out and take in what is below the knee.

12:45pm. I am just wasting time. Let me eat and do the chores."