"2:25pm. Done with breakfast and chores.
Let me get back into it. I need to figure this out.

2:30pm. My view was that the legs are too stumpy. On a whim I tried dragging the hell down, and lo and behold, the model went up a point in attractiveness.

2:35pm. Right now I am thinking what to do. No doubt it is much better now, but it is not all the way there yet. The feet are too big.

2:40pm. Hrmmm, I have no idea what is wrong. The feet are obvious, but I get that the basic shape is not all the way there. The feminine curves are right, but it still lacks gracefullness.

3pm. I am messing with the calves and where the leg meets the foot. I shrunk the later a little and it looks less stocky now, but it is still not enough.

Right now, I am confused how to do the foot. I am going to have to watch how he does it because me mental model is too vague.

Let me take a break here first.

3:25pm. Let me resume. Let me see how he does the foot.

3:40pm. https://youtu.be/VzMAh66ofq0?t=457

Here he starts working on the foot. I got distracted by him working on the body.

3:55pm. Let me just say a few words about Bitcoin since its price action has been on my mind. Its price action over the past few months since the big break has been most impressive. This is enough for me to turn bullish on it given that it has gone above 60 and is holding there for over 2 weeks already. If I was trading I'd buy around 1/3rd here, and add another 1/3 every two weeks thereafter if is giving a strong showing. Something like 65 in two weeks and over 70 in another two.

I keep hoping for BTC to collapse so that GPU prices can finally get some relief, but it does not seem to be in the cards.

Well, it could always collapse from here on out, but if that was the case it would have slid off after the big break. BTC will go to zero at some point, but it will take a bit more than than the last break from 60 to 30 to crush it.

Forget that. Let me back to the foot.

4:20pm. https://youtu.be/VzMAh66ofq0?t=478

Fuck. He is using the smooth tool to shrink the toes, while I am trying to move vertices in edit mode like a dumbass.

4:45pm. I think I got the feet done. I am doing this mostly without looking at the reference, though my creative flow.

This is actually really good for my art skill. The only reason why I am as good at programming as I am is because I am able to enter the flow zone. And it seems I have no trouble doing that with Blender. If I dedicated myself to being a 3d artist, I could make it without no trouble.

Do - think - do - think. Art is much the same as programming. There is little difference. You have some high level model, and through the process of doing you refine it.

4:50pm. So far, the tablet still hasn't arrived. I guess it won't get here today. Yeah, the courier works from 9 to 5.

Nevermind, then it is either tomorrow or on Monday. Tomorrow most likely.

4:55pm. Let me think, what do I do next.

4:55pm. What I have now strikes me as being fairly decent. But hmmmmm...it does not trike me as being beautiful.

5pm. I do not understand why. The things are thick and everything else seems to have the right curves, but overall I keep thinking old woman. Well, it is not old at this point, I think if I sketched this as a 2d drawing and showed it to the old me, he would be suitably impressed, but it not enough.

5:10pm. Focus me, let me watch Flycat some more instead of just browsing Twitter.

https://youtu.be/VzMAh66ofq0?t=505

Compared to my foot, this is masterwork. Watching him is making me realize just how poor my own mental model is.

Right, the back of the foot, where the leg meets the foot was too thick. It is things like these why I have the impression of the body being fat. Since I've been doing it from imagination, no doubt there are various other aspects I got wrong throughout as well.

5:25pm. https://youtu.be/VzMAh66ofq0?t=541

The thing where he adjusts the shape here is magical.

https://youtu.be/VzMAh66ofq0?t=544

Ok, I am sure of it now. My hips are too flat. Flycat's are peachy, while my own sort of sag.

5:35pm. Lunch time.

5:55pm. I have no idea. I want wrong in my comprehension somewhere.

https://youtu.be/VzMAh66ofq0?t=28

Here the hips are round, but if I try to round out my own model's they start to stick out and the overall figure is ruined.

6:05pm. Ok, I get it. I am missing mass on the sides. But I am not sure if my fix was right. I get the sense that I did not get the waist right.

6:15pm. Forget it. I've been in the zone for way too long to bother doing this anymore. I am not going to get good at this in a day. Tomorrow, I'll watch Flycat's video again. I got the feeling that today I really improved my comprehension of the grab tool.

In order to grasp beauty, what I need is a complete comprehension of the body. Right now I still do not understand it well enough which is why I get confused as to what a foot looks like and why the model feels saggy. I have ways more to go. A few more months working on it frmo every angle and I think I'll grasp it.

Now that I've come this far with the lower half, I think that the next step would be to move to the full body mesh apart from the head.

6:25pm. I am getting more confident in my sculpting skills, but I still have ways to go. I haven't been doing this even for a full month yet. I think at this point I should be 3 weeks into it.

I am having a difficult time believing that what I am doing here is easier than learning to draw though. Once the tablet arrives I will be able to test that view.

There is no need to worry about the time wasted. It is not like the intelligence algos are here yet, I can take my time getting good at art. A few months of practice will completely transform my view of me lacking talent in this domain."