"9:15pm. https://arxiv.org/abs/2111.05803 Gradients are Not All You Need
> The resulting loss surface of such stochastic systems are thus “smoothed” by this stochasticity and, depending on the type of randomness, this could result in better behaved loss functions (e.g. smoother, meaning the true gradient has a smaller norm), but due to the nature of how the reparameterized gradients are estimated – taking products of sequences of state Jacobian matrices – this can result in extremely high gradient norms. In some cases, throwing out the fact that the underlying system is differentiable and using black box methods to estimate the same gradient can result in a better estimate of the exact same gradient [Parmas et al., 2018, Metz et al., 2019]!

Maybe I should have gone with genetic algos for my RL agent.

> When averaging over gradient samples we can reduce variance as 1/√N where N is the number of samples, but this quickly computationally infeasible as gradients norm can grow exponentially!

> Another common technique to train in the precesses of exploding gradients is gradient clipping Pascanu et al. [2013]. Instead of taking the true gradients, one can train using gradients clipped in some way. This has proven to be of use in a variety of domains[Kaiser and Sutskever, 2015, Merity et al., 2017, Gehring et al., 2017] but will not fix all problems. As before, this calculation of the gradient is biased. To demonstrate this, we took the same Ant policy and sweep learning rate and gradient clipping strength. We found no setting which results in positive performance and thus omitted the plot.

Grad clipping is not that different from signSGD.

https://www.reddit.com/r/MachineLearning/comments/qre2vo/r_gradients_are_not_all_you_need/

///

My opinion had always been that if you have gradients, you should obviously use them, and that black box methods and evolutionary algorithms are really for situations where you don't have a readily available gradient. But the authors make a convincing argument that it's not so simple. The intuition behind Figure 1 makes the point really clear.

I really like this sort of paper that gives you a new fundamental insight!

///

Maybe I made a mistake not tossing them out outright. I admit, just doing that never occured to me. But ultimately no, that would have been a too desperate of a move for me to stomach. I can find ways of spending my time.

It should be possible to go beyond backprop rather than backwards into evolutionary algorithms.

Maybe we have it wrong. Maybe we have to come up with the right model, and then specify an optimization procedure specifically for that model. If you have a matrix multiply composed with some nonlinearity, of course then backprop is the only tool for the job. But there should be something more principled rather than random math ops.

If I only had some hint as to what that could be."