"11:45am. I am up, and why is the internet down again? At any rate, I slept quite well.
11:55am. Messing with it seems to have fixed it.

12:10pm. Let me have breakfast and do the chores first. I'll put in some overtime today to make up for missing the morning session.

1:30pm. Finally done with chores. Sheesh, today it went on forever.

2pm. Done with breakfast.

2:05pm. Let me start. Today I am thinking that I should do my work in the Illustrator. I feel that I am over reaching in trying to get good. I should just master the basic style and get on with it.

https://www.ctrlpaint.com/videos/lets-paint-a-tire-pt1

Let me watch this and then I will try to immitate it.

https://youtu.be/GRxWee6RHQM?t=263

How did he manage this? It is behind the two circles in front, but in front of the one that should be last.

Edit: Ah, he probably used one of the layers as a clipping mask. Right. Can't forge that.

2:20pm. Let me give it a try. Nevermind the graining rendering, I'll do it normally in the illustrator.

2:35pm. I am giving it a try. The gradient tool is a massive pain in the ass to work with. It would be one thing if it were like the curve tool, but this is just spew.

No I need the painting capabilities of raster software after all...

Is what I want to say, but let me try pushing it a bit further. I should at least try to do one thing to completion.

https://youtu.be/1gA0k3qiASI
How to use the Freeform Gradient Tool (THREE TRICKS) Illustrator Tutorial

Let me watch this.

https://youtu.be/1gA0k3qiASI?t=225

Yeah, it really has been bothering me how I have to change everything individually.

https://youtu.be/1gA0k3qiASI?t=498

Actually, this video he is recommending seems cool.

https://youtu.be/i55bgr5mxVo?t=6

No, this is not good enough. This is not the look I am going for. I am going to work in CSP after all.

Let me try again.

Hmmm, what is a selection layer? Ah, it takes a selection and turns it into an layer.

3:05pm. I just figured out something interesting. Fill is sensitive to opacity.

Wait, wait wait. Quick mask is in fact a regular layer than then gets turned into an selection.

That means it is possible to do the thing that I thought was impossible yesterday which is taking regular layer data and turning it into a mask.

Are quick masks the same as selection layers?

3:20pm. Oh my god, I had no idea that Clip Studio selections were so advanced. Yesterday I figured out that I can just control + click on the layer thumbnail to get a selection of the whole layer, but I noticed that this would not put the selection border around the less opaque areas. I completely thought of selection as a binary thing, but in CSP it is completely opacity sensitive. Selections and masks are always real valued. It is really a trivial thing to convert a regular layer to a quick mask/selection layer and vice versa.

Not only that, but fills are also opacity sensitive. So are deletions.

3:35pm. This is going to be super useful, especially for doing those tire patterns. I was really wondering how I was going to deal with that. Does CSP have something similar to the wrap tool. I'll have to look into it.

https://ask.clip-studio.com/en-us/detail?id=18721
Is there a way to warp a selected area (like in Photoshop)?

http://www.clip-studio.com/site/gd_en/csp/userguide/csp_userguide/500_menu/500_menu_edit_transformation_mesh.htm

4:45pm. Aghhhh, now that tired shape is giving me trouble. I do not know how to mirror and then connect it exactly. Shit. Copying it is a huge pain in the ass. Instead of using the inbuilt thing I'd have been better off just copying the entire layer.

https://youtu.be/2b4ie_FjrVc?t=66

Eyeballing the thing on the picture gave me a significantly different mental model than what I see here. How did he mirror it?

Oh, obviously there should be a way of flipping the layer.

https://www.reddit.com/r/ClipStudio/comments/e2n78k/simple_aspect_i_cant_figure_out_how_to_flip_a/
> Lasso figure, then go to Edit-->Transform-->Flip Horizontally/Flip Vertically.

You can add shortcut to those commands to your Selection Launcher Settings and/or the Command Bar Settings so you don't have to go through the menu.

Let me take a break here.

5:10pm. Let me resume. Let me repeat the drawing. The way I am doing it is wrong.

5:25pm. What I am trying to do here is turning out much harder than I thought it would.

All I am trying to do is carve out that shape so I get a smooth gap. I thought that the vector eraser would be what I need, but it giving me weird results.

5:45pm. GGGGGGGGGGAAAAAAAHHHHHH!!!

https://youtu.be/2b4ie_FjrVc?t=63

Trying to do this shit with CSP's tools is impossible. I've been trying to draw that simple shape for what must be over an hour. Shit.

6:10pm. In the end, I sketched it out using the pen and then connected it with vectors. That worked.

Right now my rage gauge is at 8/10. And trying to figure out how to make the pattern repeat is pushing it even higher.

Seriously fuck vectors. Sketching should be done the regular way after all. Vectors are just for fine tuning the lines.

6:10pm. That the internet is slow and wonky is just adding to my frustration.

6:15pm. I am trying to find a quick tip for how to space out an object and all I get is 20m videos. The Ctrl Paint guy does it in 5 seconds. I've wasted so much time on shit today.

6:20pm. https://tips.clip-studio.com/en-us/articles/2870

Making progress.

Oh, it is good I found how to activate the grid. That should improve my coordination.

6:35pm. Lunch time.

6:50pm. Let me resume.

6:55pm. Damn it, I can't get the materials to work properly. Sigh, this wasn't what I had in mind. I really should have just gone with Photoshop.

Can't I just do a copy move operation. I fucking hate how it always copies on top of the other object.

7:10pm. No an auto action would not work.

You know what, I hadn't signed up for this. I was ready to get all sorts of complicated art done, but instead following these 5m tutorials is taking me hours and hours because of the slight differences in how Photoshop and CSP work. I meant to do overtime, but all I am doing is filling up my rage meter.

I need to take a step back and recenter myself.

Today I made a definite gain since I improved my understanding of masks significantly, but I should just up down and make up my mind to understand material properties.

Right now I am thinking about my sketching attempts. I do not understand why I got so hung up on it. Using the raster pen worked better, but that was after over an hour of mucking about.

But I tried using elipses, cutting them up, using rectangles and cutting thme, but for some reason I just could not get the creative flow that way.

That is why the whole thing is so confusing. This is a very simple shape and despite that I completely faield at getting into it.

This is the kind of task that should have taken me a 5m at most.

7:25pm. I must not fall into despair or anger. Instead I have to admit that getting through the learn to draw phase might take me a whole year. If so, that is fine. I need to make up my mind to let it take however much it should take.

I need to learn from these mistakes. Tomorrow, I am going to firmly explore how to create these tiling patterns in CSP. I will master the use of materials thoroughly, do this shitty tire and then move on to the next step.

7:30pm. Let me close here for the day. I can't believe I did not get this done. Today just couldn't have gone worse."