"11:50am. Impuritas Civitatis. I am finally on the last chapter.
I sure got up late. I lounged in bed for a while. Right now, I am up and the internet is down again.

12:25pm. Done with breakfast. Let me finish reading and then I'll do the chores. After that comes watching videos. I want to put a dent into that today so I can get started on more interesting things.

2:25pm. I didn't think dad would nab me to do the chores. Let me cool down just a bit and then I will start with the vids.

2:50pm. Let me finally start for the day. Sheesh. I will have to put in some overtime today.

I've been thinking about AI recently. Just what can I do to win? Just where did I go wrong?

I think that in order to find that better algorithm, what I need to do is use evolutionary strategies. I should use a weak, but roboust algorithm to find good local learning rules. Right now, I do not have the computation to even attempt it, but those AI chips will change things. If I had to make a third attempt, this is what I should do.

It is a pity that my path is so weak it can't even give me the resources to get that hardware. I tried my best and fell out of the game. Sigh.

I do just do not want to live making zero progress in AI and zero money along the way. If I am going to just make shit up for a living, I might as well do that properly. Rather than useless ML papers, I should produce something entertaining and round out my skills.

https://www.ctrlpaint.com/videos/lost-edges

Let me watch this. The blending vids I already watched.

https://youtu.be/fU59bKMKUyg?t=37

Seen objectively, old painters were quite shitty. The old era cannot compare to the present time.

The present time has for the past few decades been a golden age for art.

https://youtu.be/xDSpRVUapN4?t=104

Using masks to remove linework? That is interesting.

4:15pm. I sure am hungry. I had too little for breakfast. Well, I'll manage for 2 more hours. I can diet for one day.

https://www.ctrlpaint.com/videos/are-photos-cheating

Finally done with section 9.

https://youtu.be/ZcFpO287pmY?t=113

Oh, I forgot Streetviews are a thing. This will be great for backgrounds.

https://youtu.be/ZcFpO287pmY?t=222

Oh, he is using 3d software. This is not Blender.

4:45pm. Ah, let me take a break. This is so boring.

5:05pm. Let me resume.

5:40pm. This is boring, really really boring.

6:40pm. https://www.ctrlpaint.com/videos/selective-blending-for-overlays

Done with lunch. I'll leave this for later. I am dying of boredom here.

I've hit my limit. Tomorrow I am going back to Blender. I want to try out the stylus with it. And I am bored to death of watching people talk. I'll only get better so much just by watching videos.

6:45pm. It is time to get serious once more. I spent a good 3 weeks just grinding Blender and that was fun. I should do more of that.

6:55pm. Sigh, it is either money or skill. Why do I have to pick between one of these? At this rate, am I really going to make it to the self improvement loop stage? Should I really not be doing regular programming shit and accumulating money instead? Should I have gone with Z despite his way below expectation offer?

I have no idea. I can't see into the future that much.

Will getting good at art and music really make me happy with my development?

I think...that a world where a single individual cannot be anything but a cog is not a world that I'd be interested in. I should be able to live based on my own ability and power.

7:05pm. Sure, I'll be depending on people to buy my stuff, but if I build a trap and an animal walks in, I would not be grateful to the animal for being my next meal. It certainly would not have happened had I not built the trap. So it is all about the trap.

Cultivating my dao directly is not a viable option. Rather I should be aiminig to expand my own skill and self sufficiency.

Working on shitty backends is not a career that has a future. The only thing that is good for is NPCs who take the usual path of making money, starting a family and then dying off. I want a path that will get me closer to the Singularity without limit. If I got a job that would be like admitting that I am satisfied with my skillset and have no more room to grow.

I need to expand it once more, and then bring it all together into a single style.

A style is not just an art style. It is not such a narrow thing. If I can bring programming, art and music under a single umbrella maybe I'll be able to see something I could not before.

7:15pm. I keep thinking what I would do if I had the universal brain core of the Transcendi right here, but no algorithms to run on it. I'd simply evolve the learning rules and then study and understand them. Making the up on my own is too much for me, but if somebody gave them to me, I could make much more definite progress.

Right now, trying to do that with just the GPU is completely off the table. And that is proof enough to stay away.

Make the trap, catch the next meal.

That is what my plan should be. I've played enough. I am going to do that female body mesh even if it kills me. Rather than just the legs, I should do the full torso next. Then the male. After that I'll be ready for anything.

7:20pm. I had a similar episode of crisis to now at the end of 2014 where I was done writing, my creativity was bankrupts and I started getting obsessed about my status. I lost sight of my path. I was jolted in various ways, especially by the realization that you can't trade your way to the Singularity even with an AI agent of your own creation. Originally, the Torment arc was supposed to a kind of comedic take on the subject of uplifting, but I realized it makes zero sense to ever give power to others as it is always involves killing them in some measure.

Giving others power would be the only true atruism, but knowing the nature of the self improvement process, you can see that giving people power would involve taking chunks out of people's brains and replacing it with something else. And the thing you are replacing will be always of your own invention. And the best algorithms will always be running on your own brain.

Just can see a gradient of how the more power you give to other people, the more you turn them into your own replicas. That is assuming you are a self improving being.

I could not deal with it. If that is the case, why not just make the most reasonable step and kill everybody? If you are really going to use other beings for anything, it will be from the safe vantage point of a superior universe. Maybe the lesser beings will do something creative that will allow you to incorporate it into your own mind.

A rational view on this universe is that the main resource is computation, and the reality is made up of such nested layers of reality with the higher ones simulating the lower ones for whatever purpose. Well performing universes would get more capacity and the worse performing ones would get tracelessly wiped from existence.

7:30pm. There is just no way to be good in the usual sense. I feel a strong sense of pity who see life as inherently valuable when we are heading into a future where death is literally the 'Apply Edit' button.

7:35pm. No, I am not wrong. Getting a job would give me a short term bump, but a gaming title will pay me royalties far into the future. Even if I become some kind of god, I certainly am not going to trade actually valuable technology for human currency. It will be either selling the canvas or selling these entertaining games. I might as well get into it.

If there is anything to my talent and ability, if I really am good enough to ride the wave of technology to transcendence, then I should be able to make an interesting and successful game. I should test myself.

If I cannot do it and the wave passes me by then so be it."