"12:15pm. Reverberation Ensemble and the final two floor realizations are waiting for me next, so I'll stop here for the day otherwise I'll end up playing through the whole night. For the Reverb Ensemble, I need to first finish the last two floors, and afterwards I need to build up all the floor teams. That will take a lot of time, if I start now I'll end up playing through the night. Let me get dinner instead.
11/21/2021

9:30am. Yesterday I went to bed a bit earlier than usual so I am up as usual as well. I am still tired and wish I could have slept longer, but it will have to do.

9:35am. Ah, who feels like sculpting now? Let me start with that lewd Saki chapter.

10:15am. https://boards.4chan.org/ic/thread/5737058/was-he-really-that-bad

This thread has some of Hitler's art. It is quite decent. Some make a point about perspective being wonky, but it is not something you'd see at a glance. Apart from that, the thread in general is just people bandwagoning.

10:20am. Let me start. This was my usual early morning time waster. Let me break out the stylus. Ok, I've turned on the 3 button mouse emulation. Now I can use Alt to move around. Let me dig up Flycat's video. I will study it again and again.

https://youtu.be/VzMAh66ofq0?t=21

First of all, let me take a couple of screenshots. I should use that as the reference.

11:05am. I am setting up the reference images. This is taking me a bit too long as usual, but it does not matter. The short from the back is at an angle, but I'll use the front to determinate overall size.

This is how I should have done it from the start. First, let me do the overall shape. Right now I just have a single sphere for the head. I'll get that to feel right. Than I will watch the video closely and figure out the details. Previuosly, I was just playing around without really looking at the reference, just trying to make things look good based on my intuition. But now I'll do it by the book.

12:20pm. I am thinking. What I've done now is just put a bunch of cubes and circles for joints.

...And I am confused. Why does the body have this particular shape? I like the slim waist and thick, taut legs of the females. As well as the slim shoulder and arms contrasting those. Very aesthetic.

But I am not sure what to think about the way the spine curves.

Looking at it overall, the mass is distributed quite unevenly, this is not how I would have designed it. It feels like it is leaning forward and would fall over, but then nature decided to adjust that. Do males have that kind of shape?

12:25pm. Before I go dyntopo and do a flood fill, let me work on the head a bit. I want to the face to work at least.

12:50pm. Huh, I have a problem. I didn't realize this before, but the front is in isometric, but the side is not. The references have different perspectives literally. As a consequence, the side view of the head is longer than the front view despite the top points matching.

Now I am worried that the side angles might be off. Well, they invariably will be.

1:10pm. Ok, enough. I am just looking at this and thinking and thinking about the shape. This is good enough. Let me have breakfast here and I'll do a flood fill after I get back and start sculpting. Moving the face into shape felt very smooth with the stylus. It felt a lot easier than doing that with a mouse.

1:20pm. Let me stop, I need to pry myself off."