"11:15am. I am up. Let me skip the morning session since it is late.
11:45am. Ok, let me watch it just for a bit. For today I am also going to get a better ref the face rather than that grainy full body one in the background.

Let me watch how Flycat does it. Rather, let me watch the retopo part.

First of all, I should absolutely make sure that my model's lips are the right size and that it is not just my imagination as to what is wrong.

It seems the PL sub monthly is still not up. Dead mod.

https://youtu.be/VzMAh66ofq0?t=789

There is a good shot here. Let me go for it.

12:05pm. Ok, the results are surprising to me. The ears, eyes, and mouth are only a tiniest bit off from the reference. Which is very good, there is no need to fix this. I though that my mouth's size might be off, but that is not it.

12:10pm. Rather the problem might be simpler. The upper lip is too far out, and that connecting bridge need just a bit more definition near the nose. I neglected to do that. Let me use the mask so I can fix this succinctly.

12:35pm. No it does not match the reference. But you know what, let me just leave it as it is. The refs lips are quite round and mine aren't.

12:40pm. Let me get breakfast. After that I'll get to work on the face. That part should not be too hard.

1:15pm. Done with chores. Since lunch will be early today, I'll wait another hour before having my meal.

Let me finally deal with the fa ce itself.

1:25pm. The clay brush really works better with the autosmoothing. This is how the clay strips is by default.

1:30pm. Good. Let me just carve the upper and lower eyelid.

2pm. Done with the eyebrows.

Let me leave for lunch.

2:25pm. I am back.

Hmmm, the eyelids look a bit too wobly for my taste. Let me max out the detail on them.

2:30pm. Hmmm, the eye wobbliness really rounds the feel.

3:05pm. Andddd...done!

Phew I am done with the base body mesh. When did I start work on this?

3:10pm. Exactly two weeks ago, a day after I started the Impuritas Civitatis chapter of Ruina. This ia long time to push a single project. But I did learn a lot about sculpting. This kind of lifestyle doing this would not be bad, but I need to get back to drawing. That having said, I should take a bunch of screenshots and post it in the /beg/ thread.

The back of the arm is a bit worrying. Let me smooth out that biceps. I do not want it to be so pronounced...no, no. Just leave it. Forget the tablet. I am done.

I should consider where to put this art. Let me take a break here.

https://www.youtube.com/results?search_query=blender+topology+rake

Let me see what topo rake does. I keep seeing that setting and do not know what it does.

https://docs.blender.org/manual/en/3.0/sculpt_paint/sculpting/tool_settings/dyntopo.html?highlight=topology%20rake

Oh there is a picture that explains it here as well. This might be a really good feature to keep on.

https://youtu.be/oWtrkSj85Ak?t=11

Ohhh, maybe I should have used face sets in order to straighten out the eyelids better. Well, things are good now so it is fine.

3:25pm. The video is useless, but the picture is worth enough on its own. It is all fine. I regret not using the clay brush with smoothing turned off more, but otherwise everything turned out quite well.

The Blender sub is quite active.

3:30pm. Ok, let me take a bunch of screenshots. I'll put them in a single file.

3:50pm. Actually forget a single file. I'll separate them out into different posts. Now let me post them on /beg/.

...No, that would end up being 19 posts. I need to marge some of them together.

4:05pm. Merged them by hand. Now let me post them.

5pm. Maybe I should have skipped posting them. But I'd be happy if even one other person saw my work. It is a bit too good to vanish into the darkness.

5:05pm. Goodbye, base mesh. I am off to bigger and better things from here on out.

I think at this point in my development I am definitely in 2/5 in art. At this rate, it will take me a full year of practice to break into 3/5, but that is fine.

5:10pm. I need to think what is next.

* Naked Luna (Hell's Prototype) cupping her breast while looking in the mirror with a grin on her face.
* Concept art of Luna in a red dress.
* Euclid (Heaven't Key) based and virtual forms.

5:15pm. You need a lot of energy and skill to do something like a comic, but an illustrated novel has lesser requirements. Good looking drawings of men and women are my next hurdle. If I can get a decent result on the above challenges, that should be enough to get started in the art department. I'll give it a month or two more.

5:25pm. Right now let me take a break. Maybe I'll call it for the day. What my immediate goal will be is to finish those Ctrl + Paint videos as well as the other tutorials. That should not take long. Then I will draw some things. I'll start with basic stuff, but I'll pick some anime images and try to replicate them without tracing. Right now I am reading the archived Symphogear thread, so I am thinking Cag or Shem-ha.

My drawing still has ways to go. In sculpting I did do that major project, so I can boast of having some skill now, but I still haven't drawn anything other than that hand sketch and the watering can. And the spatula. I need to change that. I need to take on some significant projects and push them through to completion. Right now I only have a feeling that I could do it. But like with sculpting, I have a lot to learn.

The only way to learn is to do it right.

5:35pm. I see a lot of people doing basic exercises, but I see that as a waste of time. I did not get good at programming by writing a hello world program every day. Rather the way to get good is to ramp up the complexity.

5:50pm. Bath, then I'll watch some videos. I am not done for the day yet. But every time one finishes a good project, it is good to take some relief time. After that it will be time to raise the tension again.

6:50pm. Let me watch art vids for a bit and then I'll close for the day.

7:15pm. https://www.ctrlpaint.com/videos/selective-blending-for-overlays

I've refereshed my memory of the videos I've watched last time. This is where I left off. I'll watch this tomorrow. It is time to slowly move out of the bare beginner phase in 2d art."